added withdrawal of winnings to BidderContract
