+++ scripts/helper.js
@@ -9,4 +9,15 @@ async function getCosmicGameContract() {
 	let cosmicGame = await ethers.getContractAt("CosmicGame",cosmicGameAddr)
 	return cosmicGame;
 }
-module.exports = {getCosmicGameContract};
+async function getBidderContract() {
+
+	let bidderContractAddr = process.env.BIDDER_CONTRACT_ADDRESS;
+	if ((typeof bidderContractAddr === 'undefined') || (bidderContractAddr.length != 42) )  {
+		console.log("BIDDER_CONTRACT_ADDRESS environment variable does not contain contract address");
+		process.exit(1);
+	}
+	console.log(bidderContractAddr)
+	let bidderContract = await ethers.getContractAt("BidderContract",bidderContractAddr)
+	return bidderContract;
+}
+module.exports = {getCosmicGameContract,getBidderContract};
